refactor: Add common/ directory
