+++ parser/cst/treegen.py
@@ -6,11 +6,11 @@ from .base_node import AnyNode, Node
 from .named_node import AnyNamedNode, node_from_token, node_cls_from_name
 from .nodes import *
 from .token_matcher import OpM, KwdM, Matcher, PatternT
-from ..common import region_union, RegionUnionArgT
-from ..error import BaseParseError, BaseLocatedError
+from ..common.common import region_union, RegionUnionArgT
+from ..common.error import BaseParseError, BaseLocatedError
+from ..common.str_region import StrRegion
 from ..lexer import Tokenizer
 from ..operators import UNARY_OPS, COMPARISONS, ASSIGN_OPS
-from ..str_region import StrRegion
 from ..tokens import *
 
 DT = TypeVar('DT')
